Queue scripts upgrade

+++ scripts/QueueWMC.sh
@@ -1,17 +1,24 @@
-for X in `seq 1 1 50`; do
-    scriptname="run_wmc-$1-$X.sh"
-    rm -f ${scriptname}
-    echo "#!/bin/bash" >> ${scriptname}
-    echo "#PBS -N WMC_$1-$X" >> ${scriptname}
-    echo "#PBS -l walltime=48:00:00" >> ${scriptname}
-    echo >> ${scriptname}
-    echo "cd $PWD/WMC" >> ${scriptname}
-    echo "./wmc.sh $PLUTO_OUTPUT/$1-$X.root $WMC_DATA/$1-$X.wmc.data" >> ${scriptname}
-    echo >> ${scriptname}
-    echo "rm -f $PWD/${scriptname}" >> ${scriptname}
-    chmod u+x ${scriptname}
-    qsub ${scriptname}
-    echo "${scriptname} generated and executed"
-    sleep 5
+for X in `seq 1 1 100`; do
+    if [ -e $PLUTO_OUTPUT/$1-$X.root ]; then
+	echo "number $X"
+	if [ -e $WMC_DATA/$1-$X.wmc.data ];then
+	    echo "was already done"
+	else
+	    scriptname="run_wmc-$1-$X.sh"
+	    rm -f ${scriptname}
+	    echo "#!/bin/bash" >> ${scriptname}
+	    echo "#PBS -N WMC_$1-$X" >> ${scriptname}
+	    echo "#PBS -l walltime=48:00:00" >> ${scriptname}
+	    echo >> ${scriptname}
+	    echo "cd $PWD/WMC" >> ${scriptname}
+	    echo "./wmc.sh $PLUTO_OUTPUT/$1-$X.root $WMC_DATA/$1-$X.wmc.data" >> ${scriptname}
+	    echo >> ${scriptname}
+	    echo "rm -f $PWD/${scriptname}" >> ${scriptname}
+	    chmod u+x ${scriptname}
+	    qsub ${scriptname}
+	    echo "${scriptname} generated and executed"
+	    sleep 5
+	fi
+    fi
 done
 qstat
